Move email code to emails.py
