+++ docs/code/emails.rst
@@ -1,0 +1,9 @@
+recordtransfer.emails - Email Notifications
+===========================================
+
+Each function in the ``emails.py`` file is a ``django_rq.job``, meaning that these functions are
+meant to run asynchronously with a Redis server.
+
+.. automodule:: recordtransfer.emails
+   :members:
+   :undoc-members:
